Adds test database service
Introduces a dedicated test database service with its own volume for persistence,
separating test environment data from development. This improves isolation for testing.

The test database service (database-test) was also added to the docker-compose file. Data from the dev DB is copied to the test DB before the tests are run. This ensure accurate mock data is used to run the tests.